+++ Sources/Service/Peripheral/PreparationStrategy.swift
@@ -10,8 +10,10 @@ import Foundation
 
 public protocol PreparationStrategy {
 
-    /// <#Description#>
-    /// - Parameter request: <#request description#>
-    /// - Returns: <#description#>
+    /// Handle any just-in-time transformations needed on the request before execution.
+    /// - Parameter request: The request that is about to be executed
+    /// - Returns: A modified version of the `Request` that will be executed.
+    ///
+    ///  While this method is both `async` and `throws`, any errors thrown as part of the preparation are not recoverable using any  of the`RecoveryStrategy` attached to the executing `BackendService`.
     func prepare<R>(toExecute request: Request<R>) async throws -> Request<R>
 }